fixed addition of socket-io
