+++ .idea/workspace.xml
@@ -2,15 +2,11 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="8d0cb5e6-bfb7-4fbd-859b-fefdd9f0f55e" name="Default Changelist" comment="">
+      <change afterPath="$PROJECT_DIR$/ioListeners/index.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/middleware/session.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/app.js" beforeDir="false" afterPath="$PROJECT_DIR$/app.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/database/migrations/20200614210004-create-games.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/database/migrations/20200614225114-create-locations.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/database/migrations/20200614225258-create-roles.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/database/migrations/20200614225440-create-games.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/database/migrations/20200614225646-create-players.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/database/migrations/20200614230849-create.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/database/migrations/20200614231103-create-games.js" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/index.js" beforeDir="false" afterPath="$PROJECT_DIR$/index.js" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -55,6 +51,7 @@
     <option name="RECENT_TEMPLATES">
       <list>
         <option value="CSS File" />
+        <option value="JavaScript File" />
       </list>
     </option>
   </component>
@@ -120,6 +117,22 @@
       <screen x="0" y="0" width="1440" height="900" />
     </state>
     <state x="502" y="269" key="#com.intellij.ide.scriptingContext.ui.PerFileMappingsDialogWrapper/0.0.1440.900@0.0.1440.900" timestamp="1592287918425" />
+    <state width="1807" height="278" key="GridCell.Tab.0.bottom" timestamp="1592321276494">
+      <screen x="67" y="27" width="1853" height="1053" />
+    </state>
+    <state width="1807" height="278" key="GridCell.Tab.0.bottom/67.27.1853.1053@67.27.1853.1053" timestamp="1592321276494" />
+    <state width="1807" height="278" key="GridCell.Tab.0.center" timestamp="1592321276494">
+      <screen x="67" y="27" width="1853" height="1053" />
+    </state>
+    <state width="1807" height="278" key="GridCell.Tab.0.center/67.27.1853.1053@67.27.1853.1053" timestamp="1592321276494" />
+    <state width="1807" height="278" key="GridCell.Tab.0.left" timestamp="1592321276494">
+      <screen x="67" y="27" width="1853" height="1053" />
+    </state>
+    <state width="1807" height="278" key="GridCell.Tab.0.left/67.27.1853.1053@67.27.1853.1053" timestamp="1592321276494" />
+    <state width="1807" height="278" key="GridCell.Tab.0.right" timestamp="1592321276494">
+      <screen x="67" y="27" width="1853" height="1053" />
+    </state>
+    <state width="1807" height="278" key="GridCell.Tab.0.right/67.27.1853.1053@67.27.1853.1053" timestamp="1592321276494" />
     <state width="239" height="539" key="HiddenNamespacesPopup" timestamp="1592287497485">
       <screen x="0" y="0" width="1440" height="900" />
     </state>
@@ -128,5 +141,9 @@
       <screen x="0" y="0" width="1440" height="900" />
     </state>
     <state x="266" y="23" width="997" height="797" key="SettingsEditor/0.0.1440.900@0.0.1440.900" timestamp="1592320160404" />
+    <state x="684" y="291" width="618" height="522" key="find.popup" timestamp="1592320930404">
+      <screen x="67" y="27" width="1853" height="1053" />
+    </state>
+    <state x="684" y="291" width="618" height="522" key="find.popup/67.27.1853.1053@67.27.1853.1053" timestamp="1592320930404" />
   </component>
 </project>
