+++ index.js
@@ -1,9 +1,7 @@
 require('dotenv').config();
 const cluster = require('cluster');
 const port = Number(process.env.PORT) || 3001;
-const express = require('express');
 const app = require('./app');
-const path = require('path');
 
 if (process.env.NODE_ENV === 'production') {
 	if (cluster.isMaster) {
@@ -24,11 +22,6 @@ if (process.env.NODE_ENV === 'production') {
 			cluster.fork();
 		});
 	} else {
-		app.use(express.static('./client/build'));
-		app.get('*', (req, res) =>
-			res.sendFile(path.resolve('./client/build/index.html'))
-		);
-
 		// Code to run inside of a worker
 		app.listen(port, () =>
 			console.log(`Worker ${cluster.worker.id} listening on port ${port}`)
